% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_plus.R
\name{dice_plus}
\alias{dice_plus}
\title{Diverse Clustering Ensemble Plus}
\usage{
dice_plus(...)
}
\value{
A list with the following elements
\item{E}{raw clustering ensemble object}
\item{Eknn}{clustering ensemble object with knn imputation used on \code{E}}
\item{Ecomp}{flattened ensemble object with remaining missing entries imputed
by majority voting}
\item{clusters}{final clustering assignment from the diverse clustering
ensemble method}
\item{indices}{if \code{evaluate = TRUE}, shows cluster evaluation indices;
otherwise \code{NULL}}
}
\description{
Extension of the \code{dice()} function with future enhancements,
such as feature selection and dimensionality reduction.
}
\details{
Currently, \code{dice_plus()} simply calls \code{dice()} with the same parameters.
}
\examples{
data(hgsc)
dat <- hgsc[1:100, 1:50]
ref.cl <- strsplit(rownames(dat), "_") |>
  purrr::map_chr(2) |>
  factor() |>
  as.integer()
dice.obj <- dice_plus(dat, nk = 4, reps = 5, algorithms = "hc", cons.funs =
"kmodes", ref.cl = ref.cl, progress = FALSE, verbose = FALSE)
str(dice.obj, max.level = 2)
}
\author{
Antonio C. Rodriguez
}
