% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_logger.R
\name{experiment_logger}
\alias{experiment_logger}
\title{Log an experiment result}
\usage{
experiment_logger(
  description,
  dataset,
  clustering_method,
  clustering_method_params = list(NA),
  ensemble_method = NA,
  ensemble_method_params = list(NA),
  UFS_method = "N/A",
  UFS_method_params = list(NA),
  num_features = NA,
  features = list(NA),
  dim_reduction_method = "N/A",
  dim_reduction_method_params = list(NA),
  execution_time = NA,
  labels_clustering = list(NA),
  internal_metrics = list(NA),
  external_metrics = list(NA)
)
}
\arguments{
\item{description}{Character, description of experiment.}

\item{dataset}{Character, name of the dataset.}

\item{ensemble_method}{Character, ensemble clustering method.}

\item{ensemble_method_params}{List, parameters of the ensemble clustering method.}

\item{UFS_method}{Character, unsupervised feature selection method.}

\item{UFS_method_params}{List, parameters of the unsupervised feature selection method.}

\item{num_features}{Integer, number of features selected.}

\item{features}{List, features selected.}

\item{dim_reduction_method}{Character, dimensionality reduction method.}

\item{dim_reduction_method_params}{List, parameters of the dimensionality reduction method.}

\item{execution_time}{Numeric, execution time in seconds.}

\item{labels_clustering}{ensemble clustering labels.}

\item{internal_metrics}{List, internal evaluation metrics.}

\item{external_metrics}{List, external evaluation metrics.}
}
\value{
Dataframe containing the experiment results.
}
\description{
Log an experiment result
}
