% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Gene Expression Data for Leukemia}
\format{
A data frame with 38 rows and 7,130 columns:
\describe{
\item{AFFX.BioB.5_at, AFFX.BioB.M_at, ..., Z78285_f_at}{Gene expression values (numeric).}
\item{class}{Factor indicating the leukemia subtype (\code{AML}, \code{ALL-B}, or \code{ALL-T}).}
}
}
\usage{
leukemia
}
\description{
This dataset contains gene expression profiles from patients with
leukemia. The data includes expression levels for 999 genes across 38 patient samples.
}
\details{
Each row represents a patient, while each column (except the last) corresponds
to a gene. The final column \code{class} classifies the patients as either \code{AML}
(acute myeloid leukemia), \code{ALL-B} (B-lineage acute lymphoblastic leukemia), or \code{ALL-T}
(T-lineage acute lymphoblastic leukemia).
}
\references{
Golub, T. R.; Slonim, D. K.; Tamayo, P.; Huard, C.; Gaasenbeek, M.; Mesirov, J. P.; Coller, H.; Loh, M. L.; Downing, J. R.; Caligiuri, M. A.; Bloomfield, C. D.; Lander, E. S. (1999). Molecular classification of cancer: class discovery and class prediction by gene expression monitoring. \emph{Science}, \strong{286}(5439), 531–537.
}
\keyword{datasets}
