% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_logger.R
\name{load_experiments}
\alias{load_experiments}
\title{Load experiment results from an RDS file}
\usage{
load_experiments(file = "experiments/experiment_results.rds")
}
\arguments{
\item{file}{A character string specifying the file path for retrieving experiments.
Default is "experiments/experiment_results.rds".}
}
\value{
Dataframe containing all stored experiments.
}
\description{
Load experiment results from an RDS file
}
