% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma}
\alias{lymphoma}
\title{Gene Expression Data for Lymphoma}
\format{
A data frame with 62 rows and 4027 columns:
\describe{
\item{X1, X2, ..., X4026}{Gene expression values (numeric).}
\item{class}{Factor indicating the sample type (\code{DLBCL}, \code{FL}, or \code{CLL}).}
}
}
\usage{
lymphoma
}
\description{
This dataset contains gene expression profiles from patients with
lymphoma. The data includes expression levels for 4026 genes across 62 patient samples.
}
\details{
Each row represents a patient, while each column (except the last) corresponds
to a gene. The final column \code{class} classifies the patients as either \code{DLBCL}
(Diffuse large B-cell lymphoma), \code{FL} (Follicular lymphoma), or \code{CLL} (Chronic lymphocytic leukemia).
}
\references{
Alizadeh, A. A.; Eisen, M. B.; Davis, R. E.; Ma, C.; Lossos, I. S.; Rosenwald, A.; Boldrick, J. C.; Sabet, H.; Tran, T.; Yu, X.; et al. (2000). Distinct types of diffuse large B-cell lymphoma identified by gene expression profiling. \emph{Nature}, \strong{403}(6769), 503–511.
}
\keyword{datasets}
