% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meats}
\alias{meats}
\title{Near-infrared spectra of homogenized meat samples}
\format{
A data frame with 231 rows and 1051 columns.
}
\usage{
meats
}
\description{
This dataset contains spectral measurements (400nm-2049nm) of 231 meat samples
from five species: Chicken, Beef, Turkey, Lamb, and Pork. The first 1050
columns record spectral data, while the last column indicates the sample type.
}
\references{
McElhinney, J., Downey, G., Fearn, T. (1999). Chemometric processing of visible
and near infrared reflectance spectra for species identification in selected raw homogenised meats.
\emph{Journal of Near Infrared Spectroscopy, 7}, 145–154.
}
\keyword{datasets}
