% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate_ge}
\alias{prostate_ge}
\title{Gene Expression Data for Prostate Cancer (Prostate-GE)}
\format{
A data frame with 102 rows and 5967 columns:
\describe{
\item{X1, X2, ..., X5966}{Gene expression values (numeric).}
\item{Diagnosis}{Factor indicating the sample type (\code{Tumor} or \code{Normal}).}
}
}
\usage{
prostate_ge
}
\description{
This dataset contains gene expression profiles from patients with
prostate cancer. The data includes expression levels for 5966 genes across 102 patient samples.
}
\details{
Each row represents a patient, while each column (except the last) corresponds
to a gene. The final column \code{Diagnosis} classifies the patients as either \code{Tumor} or \code{Normal}.
}
\references{
Singh, D.; Febbo, P.G.; Ross, K.; Jackson, D.G.; Manola, J.; Ladd, C.; Tamayo, P.; Renshaw, A.A.; DAmico, A.V.; Richie, J.P.; et al. (2002). Gene expression correlates of clinical prostate cancer behavior. \emph{Cancer Cell}, \strong{1}(2), 203–209.
}
\keyword{datasets}
