% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUFS.R
\name{runUFS}
\alias{runUFS}
\title{Execute Unsupervised Feature Selection (UFS) Methods in MATLAB (default parameters)}
\usage{
runUFS(
  data,
  UFS_Methods,
  ahk_executable_path = getOption("ufs.ahk_executable_path"),
  ahk_script_path = getOption("ufs.ahk_script_path")
)
}
\arguments{
\item{data}{A numeric matrix or data frame representing the dataset to be analyzed.}

\item{UFS_Methods}{A named list specifying which UFS methods should be executed.
If a method is set to \code{TRUE}, the AutoHotkey script will send \code{1} to MATLAB as input for default parameters.}

\item{ahk_executable_path}{A string specifying the path to the AutoHotkey executable (default: configured via package options).}

\item{ahk_script_path}{A string specifying the path to the AutoHotkey script (default: configured via package options).}
}
\value{
A named list containing results from each executed UFS method.
}
\description{
This function establishes a connection to a running MATLAB server, processes a given dataset
using various unsupervised feature selection (UFS) methods, retrieves results, and closes the session.

The function dynamically interacts with MATLAB, executing predefined feature selection methods while optionally
automating input handling using AutoHotkey.
}
\examples{
\dontrun{
# Define methods for feature selection
methods <- list("InfFS" = FALSE, "CFS" = FALSE, "Laplacian" = TRUE)

# Execute UFS methods with default AutoHotkey settings
results <- runUFS(Meat$x, methods)
print(results)

# Custom AutoHotkey paths
results <- runUFS(Meat$x, methods,
                  ahk_executable_path = "D:/AutoHotkey.exe",
                  ahk_script_path = "D:/send_input.ahk")
}

}
\references{
Abedinzadeh Torghabeh, F., Modaresnia, Y., Hosseini, S. A. (2023). \emph{Auto-UFSTool: An Automatic Unsupervised
Feature Selection Toolbox for MATLAB}. Journal of AI and Data Mining. doi: \href{https://doi.org/10.22044/jadm.2023.12820.2434}{10.22044/jadm.2023.12820.2434}
}
