% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUFS.R
\name{runUFS_manual_params}
\alias{runUFS_manual_params}
\title{Execute Unsupervised Feature Selection (UFS) Methods in MATLAB (manual parameters selection)}
\usage{
runUFS_manual_params(data, UFS_Methods)
}
\arguments{
\item{data}{A numeric matrix or data frame representing the dataset to be analyzed.}

\item{UFS_Methods}{A list specifying which UFS methods should be executed.}
}
\value{
A named list containing results from each executed UFS method.
}
\description{
This function establishes a connection to a running MATLAB server, processes a given dataset
using various unsupervised feature selection (UFS) methods, retrieves results, and closes the session.

The function dynamically interacts with MATLAB, executing predefined feature selection methods
}
\examples{
\dontrun{
# Define methods for feature selection
methods <- list("InfFS", "Laplacian")

# Execute UFS methods with default AutoHotkey settings
results <- runUFS(Meat$x, methods)
print(results)

}

}
\references{
Abedinzadeh Torghabeh, F., Modaresnia, Y., Hosseini, S. A. (2023). \emph{Auto-UFSTool: An Automatic Unsupervised
Feature Selection Toolbox for MATLAB}. Journal of AI and Data Mining. doi: \href{https://doi.org/10.22044/jadm.2023.12820.2434}{10.22044/jadm.2023.12820.2434}
}
