% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UFS_Methods}
\alias{UFS_Methods}
\title{Unsupervised Feature Selection Methods}
\format{
An object of class \code{list} of length 21.
}
\usage{
data(UFS_Methods)
}
\description{
List of methods used in unsupervised feature selection analysis.
If a method is set to \code{TRUE}:
the AutoHotkey script will send \code{1} to MATLAB as input for default parameters.
#' @format A list with method names and logical values indicating AutoHotkey usage.
}
\keyword{datasets}
