% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warpAR10P}
\alias{warpAR10P}
\title{warpAR10P Face Image Dataset}
\format{
A data frame with 130 rows and 2,401 columns:
\describe{
\item{X1, X2, ..., X2400}{Grayscale pixel values (numeric).}
\item{class}{Factor indicating the subject label (1–10).}
}
}
\usage{
warpAR10P
}
\description{
The warpAR10P dataset consists of grayscale face images from 10 distinct individuals. Each individual's face was imaged under varying conditions, with each image warped for alignment, resulting in high-dimensional data.
}
\details{
Each row represents an image sample, while each column (except the last) corresponds to a pixel value.
The final column \code{Person} indicates the subject's identity (from 1 to 10).
}
\references{
Li, J., Wang, K., et al. (scikit-feature). WarpAR10P available at: https://jundongl.github.io/scikit-feature/datasets.html
}
\keyword{datasets}
