% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_logger.R
\name{experiment_logger}
\alias{experiment_logger}
\title{Log an experiment result}
\usage{
experiment_logger(
  description,
  ensemble_method,
  ensemble_method_params,
  UFS_method,
  num_features,
  features,
  dim_reduction_method,
  dim_reduction_method_params,
  execution_time,
  internal_metrics,
  external_metrics
)
}
\arguments{
\item{description}{Character, description of experiment.}

\item{ensemble_method}{Character, ensemble clustering method.}

\item{ensemble_method_params}{List, parameters of the ensemble clustering method.}

\item{UFS_method}{Character, unsupervised feature selection method.}

\item{num_features}{Integer, number of features selected.}

\item{features}{List, features selected.}

\item{dim_reduction_method}{Character, dimensionality reduction method.}

\item{dim_reduction_method_params}{List, parameters of the dimensionality reduction method.}

\item{execution_time}{Numeric, execution time in seconds.}

\item{internal_metrics}{List, internal evaluation metrics.}

\item{external_metrics}{List, external evaluation metrics.}

\item{UFS_method_params}{List, parameters of the unsupervised feature selection method.}
}
\value{
Dataframe containing the experiment results.
}
\description{
Log an experiment result
}
